gen ai/requirements.txt
torch
diffusers
transformers
accelerate
Pillow
tk
